+++ src/services/query-builder.js
@@ -1,37 +1,34 @@
-'use strict';
-var _ = require('lodash');
-var Database = require('../utils/database');
+import _ from 'lodash';
+import Database from '../utils/database';
 
 function QueryBuilder(model, opts, params) {
   function hasPagination() {
     return params.page && params.page.number;
   }
 
-  this.getSkip = function () {
+  this.getSkip = () => {
     if (hasPagination()) {
-      return (parseInt(params.page.number) - 1) * this.getLimit();
-    } else {
-      return 0;
+      return (Number.parseInt(params.page.number, 10) - 1) * this.getLimit();
     }
+    return 0;
   };
 
-  this.getLimit = function () {
+  this.getLimit = () => {
     if (hasPagination()) {
-      return parseInt(params.page.size) || 10;
-    } else {
-      return 10;
+      return Number.parseInt(params.page.size, 10) || 10;
     }
+    return 10;
   };
 
-  this.getIncludes = function (modelForIncludes, fieldNamesRequested) {
-    var includes = [];
-    _.values(modelForIncludes.associations).forEach(function (association) {
+  this.getIncludes = (modelForIncludes, fieldNamesRequested) => {
+    const includes = [];
+    _.values(modelForIncludes.associations).forEach((association) => {
       if (!fieldNamesRequested ||
         (fieldNamesRequested.indexOf(association.as) !== -1)) {
         if (['HasOne', 'BelongsTo'].indexOf(association.associationType) > -1) {
           includes.push({
             model: association.target.unscoped(),
-            as: association.associationAccessor
+            as: association.associationAccessor,
           });
         }
       }
@@ -40,9 +37,9 @@ function QueryBuilder(model, opts, params) {
     return includes;
   };
 
-  this.getOrder = function (aliasName) {
+  this.getOrder = (aliasName) => {
     if (params.sort) {
-      var order = 'ASC';
+      let order = 'ASC';
 
       if (params.sort[0] === '-') {
         params.sort = params.sort.substring(1);
@@ -52,10 +49,9 @@ function QueryBuilder(model, opts, params) {
       // NOTICE: Sequelize version previous to 4.4.2 generate a bad MSSQL query
       //         if users sort the collection on the primary key, so we prevent
       //         that.
-      var idField = _.keys(model.primaryKeys)[0];
-      if (Database.isMSSQL(opts) && _.includes([idField, '-' + idField],
-        params.sort)) {
-        var sequelizeVersion = opts.sequelize.version;
+      const idField = _.keys(model.primaryKeys)[0];
+      if (Database.isMSSQL(opts) && _.includes([idField, `-${idField}`], params.sort)) {
+        const sequelizeVersion = opts.sequelize.version;
         if (sequelizeVersion !== '4.4.2-forest') {
           return null;
         }
@@ -66,9 +62,8 @@ function QueryBuilder(model, opts, params) {
         return [[opts.sequelize.col(params.sort), order]];
       } else if (aliasName) {
         return [[opts.sequelize.col(`${aliasName}.${params.sort}`), order]];
-      } else {
-        return [[params.sort, order]];
       }
+      return [[params.sort, order]];
     }
 
     return null;